Thème boostrap

+++ src/Controller/FlowController.php
@@ -5,14 +5,29 @@ namespace App\Controller;
 use App\Entity\Flow;
 use App\Form\FlowType;
 use App\Repository\FlowRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Cache\Adapter\FilesystemAdapter;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/flow')]
 class FlowController extends AbstractController
 {
+    private Security $security;
+    private FilesystemAdapter $cache;
+    private string $cacheName;
+
+    public function __construct(Security $security)
+    {
+        /** Définition du nom du cache. */
+        $this->cache = new FilesystemAdapter();
+        $this->security = $security;
+        $this->cacheName = 'cache-' . $this->security->getUser()->getUserIdentifier();
+        $this->cacheName = str_replace('@', '', $this->cacheName);
+    }
+
     #[Route('/', name: 'app_flow_index', methods: ['GET'])]
     public function index(FlowRepository $flowRepository): Response
     {
@@ -32,6 +47,9 @@ class FlowController extends AbstractController
             $flow->setUser($this->getUser());
             $flowRepository->save($flow, true);
 
+            /** Suppression du cache pour mise à jour. */
+            $this->cache->delete($this->cacheName);
+
             return $this->redirectToRoute('app_flow_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -44,6 +62,9 @@ class FlowController extends AbstractController
     #[Route('/{id}', name: 'app_flow_show', methods: ['GET'])]
     public function show(Flow $flow): Response
     {
+        // Vérifie si l'utilisateur à les droits pour voir le flux.
+        $this->denyAccessUnlessGranted('POST_VIEW', $flow);
+
         return $this->render('flow/show.html.twig', [
             'flow' => $flow,
         ]);
@@ -52,11 +73,17 @@ class FlowController extends AbstractController
     #[Route('/{id}/edit', name: 'app_flow_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Flow $flow, FlowRepository $flowRepository): Response
     {
+        // Vérifie si l'utilisateur à les droits pour éditer le flux.
+        $this->denyAccessUnlessGranted('POST_EDIT', $flow);
+
         $form = $this->createForm(FlowType::class, $flow);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $flowRepository->save($flow, true);
+
+            /** Suppression du cache pour mise à jour. */
+            $this->cache->delete($this->cacheName);
 
             return $this->redirectToRoute('app_flow_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -70,8 +97,14 @@ class FlowController extends AbstractController
     #[Route('/{id}', name: 'app_flow_delete', methods: ['POST'])]
     public function delete(Request $request, Flow $flow, FlowRepository $flowRepository): Response
     {
+        // Vérifie si l'utilisateur à les droits pour supprimer le flux.
+        $this->denyAccessUnlessGranted('POST_DELETE', $flow);
+
         if ($this->isCsrfTokenValid('delete' . $flow->getId(), $request->request->get('_token'))) {
             $flowRepository->remove($flow, true);
+
+            /** Suppression du cache pour mise à jour. */
+            $this->cache->delete($this->cacheName);
         }
 
         return $this->redirectToRoute('app_flow_index', [], Response::HTTP_SEE_OTHER);
